+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * PosixConstants: added constants that are defined in 8.3
 * [eaccess](https://www.php.net/manual/ru/function.posix-eaccess.php)
 * [fpathconf](https://www.php.net/manual/en/function.posix-fpathconf.php)
+* [sysconf](https://www.php.net/manual/en/function.posix-sysconf.php)
 
 ## 0.1.1 (15.10.2023)